cleanup eslint config, fix more linter failures
